refactor: Move and rename syncContent
