+++ dev-client/src/screens/ScreenScaffold.tsx
@@ -19,9 +19,9 @@ import {useCallback, useEffect, useState} from 'react';
 import {LayoutChangeEvent, StatusBar, StyleSheet, View} from 'react-native';
 import {SafeAreaView, useSafeAreaInsets} from 'react-native-safe-area-context';
 
-import {SyncContent} from 'terraso-mobile-client/components/buttons/SyncButton';
 import {Box, Column} from 'terraso-mobile-client/components/NativeBaseAdapters';
 import {RestrictByFlag} from 'terraso-mobile-client/components/RestrictByFlag';
+import {SyncContent} from 'terraso-mobile-client/components/SyncContent';
 import {useHeaderHeight} from 'terraso-mobile-client/hooks/useHeaderHeight';
 import {AppBar} from 'terraso-mobile-client/navigation/components/AppBar';
 import {theme} from 'terraso-mobile-client/theme';
